:smile 改用Mybatis-Plus，重构业务层。

+++ src/main/java/com/example/demo/vo/UserInfoVo.java
@@ -14,7 +14,7 @@ public class UserInfoVo {
 
     private String email;
 
-    private Integer gender;
+    private Short gender;
 
     @JsonFormat(pattern = "yyyy-MM-dd", timezone = "GMT+8")
     private Date birthday;
